Add script to download artifacts.
