+++ build_tools/buildkite/scripts/wait_for_pipeline_success.py
@@ -12,188 +12,21 @@ triggered (or pre-existing) build succeeds, otherwise fails.
 """
 
 import argparse
-import os
-from pybuildkite import buildkite
 import sys
-import time
+import json
+from typing import Optional
 
-# Fake build to return when running locally.
-FAKE_PASSED_BUILD = dict(number=42, state="passed")
-
-
-def get_build_number(build):
-  return build.get("number")
+from pybuildkite import buildkite
+from common.buildkite_pipeline_manager import BuildkitePipelineManager
+from common.buildkite_utils import BuildObject, get_build_number, get_build_state, linkify
 
 
-class BuildkitePipelineManager():
-
-  def __init__(
-      self,
-      *,
-      access_token,
-      pipeline,
-      organization,
-      commit,
-      branch,
-      message=None,
-      author_name,
-      author_email,
-      pull_request_base_branch=None,
-      pull_request_id=None,
-      pull_request_repository=None,
-  ):
-    self._buildkite = buildkite.Buildkite()
-    self._buildkite.set_access_token(access_token)
-    self._pipeline = pipeline
-    self._organization = organization
-    self._commit = commit
-    self._message = message
-    self._branch = branch
-    self._author_name = author_name
-    self._author_email = author_email
-    self._author = {"name": self._author_name, "email": self._author_email}
-    self._pull_request_base_branch = pull_request_base_branch
-    self._pull_request_id = pull_request_id
-    self._pull_request_repository = pull_request_repository
-
-  @staticmethod
-  def from_environ(pipeline):
-    # A token for the Buildkite API. Needs read/write privileges on builds to
-    # watch and create builds. Within our pipelines we fetch this from secret
-    # manager: https://cloud.google.com/secret-manager. Users can create a
-    # personal token for running this script locally:
-    # https://buildkite.com/docs/apis/managing-api-tokens
-    access_token = os.environ["BUILDKITE_ACCESS_TOKEN"]
-
-    # Buildkite sets these environment variables. See
-    # https://buildkite.com/docs/pipelines/environment-variables. If running
-    # locally you can set locally or use the simulate_buildkite.sh script.
-    organization = os.environ["BUILDKITE_ORGANIZATION_SLUG"]
-    commit = os.environ["BUILDKITE_COMMIT"]
-    branch = os.environ["BUILDKITE_BRANCH"]
-
-    # These variables aren't strictly necessary. Just nice to have (and set by
-    # Buildkite).
-    author_name = os.environ.get("BUILDKITE_BUILD_AUTHOR")
-    author_email = os.environ.get("BUILDKITE_BUILD_AUTHOR_EMAIL")
-    message = os.environ.get("BUILDKITE_MESSAGE")
-    # These may not be set if build is not from a pull request
-    pull_request_id = os.environ.get("BUILDKITE_PULL_REQUEST")
-    pull_request_base_branch = os.environ.get(
-        "BUILDKITE_PULL_REQUEST_BASE_BRANCH")
-    pull_request_repository = os.environ.get("BUILDKITE_PULL_REQUEST_REPO")
-
-    return BuildkitePipelineManager(
-        access_token=access_token,
-        organization=organization,
-        pipeline=pipeline,
-        commit=commit,
-        message=message,
-        branch=branch,
-        author_name=author_name,
-        author_email=author_email,
-        pull_request_id=pull_request_id,
-        pull_request_base_branch=pull_request_base_branch,
-        pull_request_repository=pull_request_repository,
-    )
-
-  def get_builds(self):
-    # Avoid API calls when running locally. The local organization doesn't
-    # exist.
-    if self._organization == "local":
-      return [FAKE_PASSED_BUILD]
-    return self._buildkite.builds().list_all_for_pipeline(
-        organization=self._organization,
-        pipeline=self._pipeline,
-        commit=self._commit)
-
-  def get_build_by_number(self, build_number):
-    # Avoid API calls when running locally. The local organization doesn't
-    # exist.
-    if (self._organization == "local" and
-        build_number == get_build_number(FAKE_PASSED_BUILD)):
-      print("Returning fake build because running locally")
-      return FAKE_PASSED_BUILD
-    return self._buildkite.builds().get_build_by_number(self._organization,
-                                                        self._pipeline,
-                                                        build_number)
-
-  def get_latest_build(self):
-    all_builds = self.get_builds()
-    if not all_builds:
-      return None
-    return max(all_builds, key=get_build_number)
-
-  def create_build(self):
-    return self._buildkite.builds().create_build(
-        organization=self._organization,
-        pipeline=self._pipeline,
-        commit=self._commit,
-        message=self._message,
-        branch=self._branch,
-        ignore_pipeline_branch_filters=True,
-        author=self._author,
-        pull_request_base_branch=self._pull_request_base_branch,
-        pull_request_id=self._pull_request_id,
-        pull_request_repository=self._pull_request_repository,
-    )
-
-  def wait_for_build(self, build_number):
-    # We want to override the previous output when logging about waiting, so
-    # this doesn't print a bunch of unhelpful log lines. Carriage return takes
-    # us back to the beginning of the line, but it doesn't override previous
-    # output past the end of the new output, so we pad things to ensure that
-    # each line is at least as long as the previous one. Note that this approach
-    # only works if a print statement doesn't overflow a single line (at least
-    # on my machine). In that case, the beginning of the line is partway through
-    # the previous print, although it at least starts on a new line. Better
-    # suggestions welcome.
-    min_line_length = 0
-    # We don't need great precision
-    start = time.monotonic()
-    while True:
-      build = self.get_build_by_number(build_number)
-      state = buildkite.BuildState(build["state"])
-      wait_time = int(round(time.monotonic() - start))
-      if state in [
-          buildkite.BuildState.PASSED,
-          buildkite.BuildState.FAILED,
-          buildkite.BuildState.CANCELED,
-          buildkite.BuildState.SKIPPED,
-          buildkite.BuildState.NOT_RUN,
-      ]:
-        output_str = f"Build finished in state '{state.name}'"
-        min_line_length = max(min_line_length, len(output_str))
-        print(output_str.ljust(min_line_length))
-        return state
-
-      output_str = (
-          f"Waiting for build {build_number} to complete. Waited {wait_time}"
-          f" seconds. Currently in state '{state.name}':"
-          f" {linkify(self.get_url_for_build(build_number))}")
-      min_line_length = max(min_line_length, len(output_str))
-      print(output_str.ljust(min_line_length), "\r", end="", flush=True)
-      # Yes, polling is unfortunately the best we can do here :-(
-      time.sleep(5)
-
-  def get_url_for_build(self, build_number):
-    return f"https://buildkite.com/{self._organization}/{self._pipeline}/builds/{build_number}"
-
-
-# Make a link clickable using ANSI escape sequences. See
-# https://buildkite.com/docs/pipelines/links-and-images-in-log-output
-def linkify(url, text=None):
-  if text is None:
-    text = url
-
-  return f"\033]1339;url={url};content={text}\a"
-
-
-def should_create_new_build(bk, build, rebuild_option):
+def should_create_new_build(bk: BuildkitePipelineManager,
+                            build: Optional[BuildObject], rebuild_option: str):
   if not build:
     print("Didn't find previous build for pipeline. Creating a new one.")
     return True
-  state = buildkite.BuildState(build["state"])
+  state = get_build_state(build)
   build_number = get_build_number(build)
   url = bk.get_url_for_build(build_number)
   print(f"Found previous build with state '{state.name}': {url}")
@@ -218,7 +51,7 @@ def should_create_new_build(bk, build, rebuild_option):
   return False
 
 
-def parse_args():
+def parse_args() -> argparse.Namespace:
   parser = argparse.ArgumentParser(
       description="Waits on the status of the last Buildkite build for a given"
       " commit or creates such a build if none exists")
@@ -232,6 +65,10 @@ def parse_args():
       " on build finished in state other than 'passed'",
       choices=["force", "failed", "bad"],
   )
+  parser.add_argument(
+      "--output-build-json",
+      help="Output path of the build object in JSON, which contains the details"
+      " of the successful build.")
   return parser.parse_args()
 
 
@@ -249,6 +86,11 @@ def main(args):
     print(f"Build was not successful: {linkify(url)}")
     sys.exit(1)
   print(f"Build completed successfully: {linkify(url)}")
+
+  if args.output_build_json:
+    build = bk.get_build_by_number(build_number)
+    with open(args.output_build_json, "w") as f:
+      json.dump(build, f)
 
 
 if __name__ == "__main__":
